Relax pytest version requirement
we do depend on tcms-api version but we'll try to accomodate any
pytest version so we don't force users to upgrade their
environments b/c for some teams having the latest and greatest tools
is not the preferred option.

In Travis CI we will always test with the latest available anyway.